requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
langchain-pinecone

ollama
streamlit
beautifulsoup4
pinecone-client

python-dotenv
tiktoken
langchainhub